Moving to conf system - ColorMixin 01
