+++ src/base/i-color-mixin-conf.ts
@@ -1,5 +1,6 @@
 import {IBaseMixinConf} from './i-base-mixin-conf';
+import {BaseAccessor} from '../core/types';
 
 export interface IColorMixinConf extends IBaseMixinConf {
-
+    colorCalculator?: BaseAccessor<string>;
 }
